[doc] :book: device placement configuration
1. Device placement documentation
2. Fix bug in device placement for Hysia Decoder
