+++ server/config/__init__.py
@@ -11,8 +11,8 @@ CONFIG_DIR = Path(__file__).absolute().parents[1] / 'config'
 # load config
 with open(CONFIG_DIR / 'weight_config.yml', 'r') as stream:
     try:
-        config = yaml.safe_load(stream)
-        config = dict_to_object(config)
+        weight_config = yaml.safe_load(stream)
+        weight_config = dict_to_object(weight_config)
     except yaml.YAMLError as e:
         print(e)
         exit(1)
@@ -26,4 +26,4 @@ with open(CONFIG_DIR / 'device_placement.yml') as stream:
         print(e)
         exit(1)
 
-__all__ = ['WEIGHT_DIR', 'config', 'device_config']
+__all__ = ['WEIGHT_DIR', 'weight_config', 'device_config']
