+++ server/model_server/feature_extract_model_server.py
@@ -8,13 +8,13 @@ import grpc
 import numpy as np
 from PIL import Image
 
+from config import weight_config, WEIGHT_DIR, device_config
 from hysia.dataset.srt_handler import extract_srt
 from hysia.models.nlp.sentence import TF_Sentence
 from hysia.models.object.audioset_feature_extractor import AudiosetFeatureExtractor
 from hysia.models.scene.detector import scene_visual
 from hysia.utils.logger import Logger
-from model_server import config, WEIGHT_DIR, device_config
-from model_server.misc import load_tf_graph, obtain_device
+from hysia.utils.misc import obtain_device, load_tf_graph
 from protos import api2msl_pb2, api2msl_pb2_grpc
 
 # Time constant
@@ -27,15 +27,15 @@ logger = Logger(
 
 
 def load_sentence_model():
-    model_path = str(WEIGHT_DIR / config.feature_extraction.sentence_encoder)
+    model_path = str(WEIGHT_DIR / weight_config.feature_extraction.sentence_encoder)
     # Instantiate sentence feature extractor
     return TF_Sentence(model_path)
 
 
 def load_audio_model():
     # Instantiate audio feature extractor
-    vgg_graph = load_tf_graph(WEIGHT_DIR / config.feature_extraction.vggish.fr)
-    vgg_pca_path = WEIGHT_DIR / config.feature_extraction.vggish.pca
+    vgg_graph = load_tf_graph(WEIGHT_DIR / weight_config.feature_extraction.vggish.fr)
+    vgg_pca_path = WEIGHT_DIR / weight_config.feature_extraction.vggish.pca
 
     audio_model = AudiosetFeatureExtractor(vgg_graph, vgg_pca_path)
     return audio_model
@@ -59,7 +59,7 @@ class Api2MslServicer(api2msl_pb2_grpc.Api2MslServicer):
         self.audio_model = load_audio_model()
 
         # load object detection model
-        places365_config = config.scene.places365
+        places365_config = weight_config.scene.places365
         self.backbone = places365_config.backbone
         model_path = str(WEIGHT_DIR / places365_config.model)
         label_path = WEIGHT_DIR / places365_config.label
