+++ server/model_server/product_search_model_server.py
@@ -6,14 +6,14 @@ from concurrent import futures
 # rpc imports
 import grpc
 
+from config import device_config
 from hysia.search.product_search import ProductSearch
 from hysia.utils.logger import Logger
 from hysia.utils.perf import StreamSuppressor
-from model_server import device_config
-from model_server.misc import obtain_device
 from protos import api2msl_pb2, api2msl_pb2_grpc
-
 # Time constant
+from utils.misc import obtain_device
+
 _ONE_DAY_IN_SECONDS = 24 * 60 * 60
 
 VIDEO_DATA_PATH = '/data/disk2/hysia_data/Stanford_Online_Products/'
