+++ server/model_server/scene_search_model_server.py
@@ -6,14 +6,13 @@ from pathlib import Path
 
 import grpc
 
+from config import device_config
 from hysia.search.search import DatabasePklSearch
 from hysia.utils.logger import Logger
 from hysia.utils.perf import StreamSuppressor
-from model_server import device_config
-from model_server.misc import obtain_device
 from protos import api2msl_pb2, api2msl_pb2_grpc
+from utils.misc import obtain_device
 
-# Time constant
 _ONE_DAY_IN_SECONDS = 24 * 60 * 60
 
 logger = Logger(
